requirements.txt
uvicorn==0.34.2
fastapi==0.115.12
pydantic==2.11.4
pydantic-settings==2.9.1
sqlalchemy==2.0.41
python-jose==3.4.0
passlib==1.7.4
aiosqlite==0.21.0
pydantic[email]==2.11.4
python-multipart==0.0.20
httpx==0.28.1